+++ app/src/main/java/com/owoshopkeeperpanel/Model/Owo_product.java
@@ -3,7 +3,7 @@ package com.owoshopkeeperpanel.Model;
 import java.io.Serializable;
 
 public class Owo_product implements Serializable {
-    private int product_id;
+    private long product_id;
     private String product_name;
     private String product_category;
     private double product_price;
@@ -34,11 +34,11 @@ public class Owo_product implements Serializable {
         this.product_image = product_image;
     }
 
-    public int getProduct_id() {
+    public long getProduct_id() {
         return product_id;
     }
 
-    public void setProduct_id(int product_id) {
+    public void setProduct_id(long product_id) {
         this.product_id = product_id;
     }
