+++ app/src/main/java/com/owoshopkeeperpanel/Model/product_database_model.java
@@ -1,24 +1,24 @@
 package com.owoshopkeeperpanel.Model;
 
 public class product_database_model {
-    int product_id;
+    long product_id;
     String product_quantity;
     String product_price;
 
     public product_database_model() {
     }
 
-    public product_database_model(int product_id, String product_quantity, String product_price) {
+    public product_database_model(long product_id, String product_quantity, String product_price) {
         this.product_id = product_id;
         this.product_quantity = product_quantity;
         this.product_price = product_price;
     }
 
-    public int getProduct_id() {
+    public long getProduct_id() {
         return product_id;
     }
 
-    public void setProduct_id(int product_id) {
+    public void setProduct_id(long product_id) {
         this.product_id = product_id;
     }
